+++ constcheck/_typing.py
@@ -17,5 +17,11 @@ PathLike = _t.Union[str, _os.PathLike]
 TokenList = _t.List[_TokenText]
 ValueTuple = _t.Tuple[int, int]
 ArgTuple = _t.Tuple[
-    PathLike, ValueTuple, bool, bool, _t.Optional[str], _t.List[str]
+    PathLike,
+    ValueTuple,
+    bool,
+    bool,
+    _t.Optional[str],
+    _t.List[str],
+    _t.List[str],
 ]